requirements.txt
sigen~=0.1.9
beautifulsoup4~=4.13.4
requests~=2.32.3
python-dotenv~=1.0.1
openpyxl~=3.1.5